requirements.txt
langchain==0.2.14
# langchain
langchain-community==0.2.12
# langchain-community
langchain-groq===0.1.9
# langchain-groq
unstructured==0.14.6
# unstructured
selenium==4.21.0
# selenium
# chromadb==0.5.0
# chromadb==0.4.0
# chromadb
# streamlit==1.35.0
streamlit
pandas==2.0.2
python-dotenv==1.0.0